Omair commits/changes..

+++ src/navigation/screens/CompanyProfile.js
@@ -1,7 +1,7 @@
-import styles from "./profile.module.css"
+import styles from "./profile.module.css";
 import { FaPenToSquare } from "react-icons/fa6";
 import { Suspense, useEffect, useState } from "react";
-import { redirect, useLoaderData, useNavigate } from "react-router-dom";
+import { Link, redirect, useLoaderData, useNavigate } from "react-router-dom";
 import DropDown from "../../components/DropDown";
 import "react-dropdown/style.css";
 import {
@@ -22,9 +22,9 @@ export const loader = async () => {
 
 const CompanyProfile = () => {
   const { company, uType, countries } = useLoaderData();
-  console.log(company);
+  // console.log(company);
 
-  const [loading, setLoading] = useState(true);
+  const [loading, setLoading] = useState(false);
   const [companyData, setCompanyData] = useState(company);
   const [Image, setImage] = useState({ file: null, url: null });
   const [dataImage, setDataImage] = useState({ file: null, url: null });
@@ -32,6 +32,7 @@ const CompanyProfile = () => {
   const [cities, setCities] = useState([{}]);
   const [country, setCountry] = useState(company?.companycountry);
   const [city, setCity] = useState(company?.companycity);
+  const [emailExist, setEmailExist] = useState(false);
   const [companyCategory, setCompanyCategory] = useState(
     CompanyCategories.filter((con) => con.index === company?.categoryid)[0]
       ?.value
@@ -63,47 +64,68 @@ const CompanyProfile = () => {
     try {
       const localStrData = JSON.parse(localStorage.getItem("user"));
       const type = localStrData.userType;
-      const response = await fetch("api/company/companyDetails", {
-        method: "get",
-        headers: {
-          "Content-Type": "application/json",
-          Authorization: "Bearer " + localStrData.token,
-          type: type,
-        },
-      });
+      const response = await fetch(
+        "https://work-it-back.vercel.app/api/company/companyDetails",
+        {
+          method: "get",
+          headers: {
+            "Content-Type": "application/json",
+            Authorization: "Bearer " + localStrData.token,
+            type: type,
+          },
+        }
+      );
       const json = await response.json();
       setCompanyData(json);
     } catch (error) {
       console.error("Error fetching company data:", error);
-    } finally {
-      setLoading(false);
     }
   };
 
   const handelSubmit = async () => {
+    setLoading(true);
+    let imageUrl;
     try {
-      if (Image.url != null) handleFileChange();
+      if (Image.url != null) {
+        imageUrl = await handleFileChange();
+      }
 
       const localStrData = JSON.parse(localStorage.getItem("user"));
       const type = localStrData.userType;
       localStrData.user = companyData[0];
       localStrData.user.country = country;
       localStrData.user.city = city;
+      localStrData.user.imageUrl = imageUrl ? imageUrl : companyData[0].logo;
 
-      const response = await fetch("/api/company/editProfile", {
-        method: "put",
-        headers: {
-          "Content-Type": "application/json",
-          authorization: "token: " + localStrData.token,
-          type: type,
-        },
-        body: JSON.stringify({ companyData: localStrData.user }),
-      });
-      const updatedUserDataString = JSON.stringify(localStrData);
-      localStorage.setItem("user", updatedUserDataString);
-      navigate("/companyview");
+      const response = await fetch(
+        "https://work-it-back.vercel.app/api/company/editProfile",
+        // "http://localhost:5002/api/company/editProfile",
+
+        {
+          method: "put",
+          headers: {
+            "Content-Type": "application/json",
+            authorization: "token: " + localStrData.token,
+            type: type,
+          },
+          body: JSON.stringify({ companyData: localStrData.user }),
+        }
+      );
+
+      const json = await response.json();
+
+      if (json.error == "Email already exists") {
+        setEmailExist(true);
+      } else {
+        setEmailExist(false);
+        const updatedUserDataString = JSON.stringify(localStrData);
+        localStorage.setItem("user", updatedUserDataString);
+        navigate("/companyview");
+      }
     } catch (error) {
       console.error("Error:", error);
+    } finally {
+      setLoading(false);
     }
   };
 
@@ -111,12 +133,15 @@ const CompanyProfile = () => {
     const localStrData = JSON.parse(localStorage.getItem("user"));
     const type = localStrData.userType;
     try {
-      const response = await fetch("/api/company/gallery", {
-        headers: {
-          authorization: "token: " + localStrData.token,
-          type: type,
-        },
-      });
+      const response = await fetch(
+        "https://work-it-back.vercel.app/api/company/gallery",
+        {
+          headers: {
+            authorization: "token: " + localStrData.token,
+            type: type,
+          },
+        }
+      );
       const data = await response.json();
 
       if (data.success) {
@@ -141,28 +166,32 @@ const CompanyProfile = () => {
 
   const handleFileChange = async () => {
     try {
-      const localStrData = JSON.parse(localStorage.getItem("user"));
-      const type = localStrData.userType;
-      const formData = new FormData();
-      formData.append("image", dataImage);
-
-      const response = await fetch("/api/company/uploadImage", {
-        method: "POST",
-        headers: {
-          authorization: "token: " + localStrData.token,
-          type: type,
-        },
-        body: formData,
-      });
-
-      if (response.ok) {
-        const result = await response.json();
-        console.log("Image uploaded successfully:", result.imageUrl);
+      let imageURL;
+      if (
+        dataImage &&
+        (dataImage.type === "image/png" ||
+          dataImage.type === "image/jpg" ||
+          dataImage.type === "image/jpeg")
+      ) {
+        const image = new FormData();
+        image.append("file", dataImage);
+        image.append("cloud_name", "dapnyieo6");
+        image.append("upload_preset", "gyeiwufc");
+        const response = await fetch(
+          "https://api.cloudinary.com/v1_1/dapnyieo6/image/upload",
+          {
+            method: "post",
+            body: image,
+          }
+        );
+        const imgData = await response.json();
+        imageURL = imgData.url.toString();
+        return imageURL;
       } else {
-        console.error("Failed to upload image");
+        alert("type should be png or jpg or jpeg ");
       }
     } catch (error) {
-      console.error("Error:", error);
+      console.error("Error while uploading image:", error);
     }
   };
   const refrechCities = async (country) => {
@@ -182,7 +211,7 @@ const CompanyProfile = () => {
     ]);
   }, [companyCategory]);
 
-  if (loading) return <Loader />;
+  // if (loading) return <Loader />;
 
   return (
     <div>
@@ -190,158 +219,178 @@ const CompanyProfile = () => {
       <h2 className={styles.text} style={{ textAlign: "center" }}>
         EDIT YOUR COMPANY PROFILE
       </h2>
-      <div className={styles.main}>
-        <div className={styles.user}>
-          <div
-            style={{
-              width: "150px",
-              height: "150px",
-              overflow: "hidden",
-              borderRadius: "50%",
-            }}
-            className={styles.photo}
-          >
-            {Image.url && (
-              <img
-                src={Image.url}
-                alt="Profile"
-                style={{
-                  width: "100%",
-                  height: "100%",
-                  objectFit: "cover",
-                  borderRadius: "50%",
-                }}
-              />
-            )}
-            {!Image.url && (
-              <img
-                src={profilePhoto.logo}
-                alt="Profile"
-                style={{
-                  width: "100%",
-                  height: "100%",
-                  objectFit: "cover",
-                  borderRadius: "50%",
-                }}
-              />
-            )}
-          </div>
-          <label
-            htmlFor="fileInput"
-            className="file-input-label"
-            style={{
-              position: "relative",
-              overflow: "hidden",
-              width: "fit-content",
-            }}
-          >
-            <FaPenToSquare className="pen-icon" style={{ cursor: "pointer" }} />
-            <input
-              type="file"
-              id="fileInput"
-              className="file-input"
+      {loading ? (
+        <Loader />
+      ) : (
+        <div className={styles.main}>
+          <div className={styles.user}>
+            <div
               style={{
-                position: "absolute",
-                opacity: 0,
-                top: 0,
-                left: 0,
-                width: "100%",
-                height: "100%",
-                cursor: "pointer",
+                width: "150px",
+                height: "150px",
+                overflow: "hidden",
+                borderRadius: "50%",
               }}
-              onChange={(e) => handleImageChange(e)}
+              className={styles.photo}
+            >
+              {Image.url && (
+                <img
+                  src={Image.url}
+                  alt="Profile"
+                  style={{
+                    width: "100%",
+                    height: "100%",
+                    objectFit: "cover",
+                    borderRadius: "50%",
+                  }}
+                />
+              )}
+              {!Image.url && (
+                <img
+                  src={profilePhoto.logo}
+                  alt="Profile"
+                  style={{
+                    width: "100%",
+                    height: "100%",
+                    objectFit: "cover",
+                    borderRadius: "50%",
+                  }}
+                />
+              )}
+            </div>
+            <label
+              htmlFor="fileInput"
+              className="file-input-label"
+              style={{
+                position: "relative",
+                overflow: "hidden",
+                width: "fit-content",
+              }}
+            >
+              <FaPenToSquare
+                className="pen-icon"
+                style={{ cursor: "pointer" }}
+              />
+              <input
+                type="file"
+                id="fileInput"
+                accept="image/png, image/jpeg"
+                className="file-input"
+                style={{
+                  position: "absolute",
+                  opacity: 0,
+                  top: 0,
+                  left: 0,
+                  width: "100%",
+                  height: "100%",
+                  cursor: "pointer",
+                }}
+                onChange={(e) => handleImageChange(e)}
+              />
+            </label>
+          </div>
+          <div className={styles.first_name}>
+            <span className={styles.text}>Name</span>
+            <input
+              type="text"
+              placeholder="Company Name"
+              defaultValue={companyData[0]?.companyname}
+              onChange={(e) =>
+                setCompanyData((prevUserData) => [
+                  { ...prevUserData[0], companyname: e.target.value },
+                ])
+              }
             />
-          </label>
-        </div>
-        <div className={styles.first_name}>
-          <span className={styles.text}>Name</span>
-          <input
-            type="text"
-            placeholder="Company Name"
-            defaultValue={companyData[0]?.companyname}
-            onChange={(e) =>
-              setCompanyData((prevUserData) => [
-                { ...prevUserData[0], companyname: e.target.value },
-              ])
-            }
-          />
-        </div>
-        <div className={styles.last_name}>
-          <span className={styles.text}>Company size</span>
-          {/* [x] from sizes, do the same with categories */}
-          <select
-            placeholder="Company Size"
-            className="text_input"
-            style={{
-              border: "1px solid rgb(204, 204, 204)",
-              width: "fit-content",
-            }}
-            defaultValue={companyData[0]?.company_size}
-            onChange={(e) =>
-              setCompanyData((prevUserData) => [
-                { ...prevUserData[0], company_size: e.target.value },
-              ])
-            }
-          >
-            {CompanySizes?.map((s) => (
-              <option key={s.index} value={s.value}>
-                {s.value}
-              </option>
-            ))}
-          </select>
-        </div>
-        <div className={styles.email}>
-          <span className={styles.text}>Email</span>
-          <input
-            type="text"
-            placeholder="Company Email"
-            value={companyData[0]?.companyemail}
-            onChange={(e) =>
-              setCompanyData((prevUserData) => [
-                { ...prevUserData[0], companyemail: e.target.value },
-              ])
-            }
-          />
-        </div>
-        <div className={styles.pass}>
-          <span className={styles.text}>Password</span>
-          <input type="password" placeholder="**********" disabled />
-        </div>
-        <div className={styles.phone}>
-          <span className={styles.text}>Company category</span>
-          <DropDown
-            data={CompanyCategories}
-            state={companyCategory}
-            setState={setCompanyCategory}
-            placeholder="Company category"
-            className="text_input"
-          />
-        </div>
-        <div className={styles.address}>
-          <span className={styles.text}>Address</span>
-          <div className={styles.address_inputs}>
-            <Suspense fallback={<Loader />}>
-              <DropDown
-                data={countries}
-                placeholder={company?.companycountry}
-                className="text_input"
-                state={country}
-                setState={setCountry}
-              />
-              <DropDown
-                data={cities}
-                placeholder={company?.companycity}
-                className="text_input"
-                state={city}
-                setState={setCity}
-              />
-            </Suspense>
+          </div>
+          <div className={styles.last_name}>
+            <span className={styles.text}>Company size</span>
+            {/* [x] from sizes, do the same with categories */}
+            <select
+              placeholder="Company Size"
+              className="text_input"
+              style={{
+                border: "1px solid rgb(204, 204, 204)",
+                width: "fit-content",
+              }}
+              defaultValue={companyData[0]?.company_size}
+              onChange={(e) =>
+                setCompanyData((prevUserData) => [
+                  { ...prevUserData[0], company_size: e.target.value },
+                ])
+              }
+            >
+              {CompanySizes?.map((s) => (
+                <option key={s.index} value={s.value}>
+                  {s.value}
+                </option>
+              ))}
+            </select>
+          </div>
+          <div className={styles.email}>
+            <span className={styles.text}>Email</span>
+            <input
+              type="text"
+              placeholder="Company Email"
+              value={companyData[0]?.companyemail}
+              onChange={(e) =>
+                setCompanyData((prevUserData) => [
+                  { ...prevUserData[0], companyemail: e.target.value },
+                ])
+              }
+            />
+            {emailExist ? <p className="error">Email already exists</p> : <></>}
+          </div>
+          <div className={styles.pass}>
+            <span className={styles.text}>Password</span>
+            <Link
+              to="/changePassword"
+              state={{ oldPassword: companyData[0]?.password, type: "company" }}
+              className={styles.save}
+              style={{
+                justifyContent: "normal",
+                marginBottom: 0,
+                textDecoration: "none",
+              }}
+            >
+              <button>Change</button>
+            </Link>
+          </div>
+          <div className={styles.phone}>
+            <span className={styles.text}>Company category</span>
+            <DropDown
+              data={CompanyCategories}
+              state={companyCategory}
+              setState={setCompanyCategory}
+              placeholder="Company category"
+              className="text_input"
+            />
+          </div>
+          <div className={styles.address}>
+            <span className={styles.text}>Address</span>
+            <div className={styles.address_inputs}>
+              <Suspense fallback={<Loader />}>
+                <DropDown
+                  data={countries}
+                  placeholder={company?.companycountry}
+                  className="text_input"
+                  state={country}
+                  setState={setCountry}
+                />
+                <DropDown
+                  data={cities}
+                  placeholder={company?.companycity}
+                  className="text_input"
+                  state={city}
+                  setState={setCity}
+                />
+              </Suspense>
+            </div>
+          </div>
+          <div className={styles.save}>
+            <button onClick={handelSubmit}>Save</button>
           </div>
         </div>
-        <div className={styles.save}>
-          <button onClick={handelSubmit}>Save</button>
-        </div>
-      </div>
+      )}
     </div>
   );
 };
